+++ src/handlers/mod.rs
@@ -1,0 +1,2 @@
+pub mod plans;
+pub mod readings;
