+++ src/lib.rs
@@ -1,0 +1,4 @@
+pub mod datastore;
+pub mod handlers;
+pub mod models;
+pub mod routes;
