+++ src/main.rs
@@ -1,10 +1,12 @@
 mod datastore;
-mod http;
+mod handlers;
+mod models;
+mod routes;
 
 #[tokio::main]
 async fn main() {
-    // build our application with a single route
-    let app = http::build().await;
+    // Build our application with a single route
+    let app = routes::build().await;
 
     let listener = tokio::net::TcpListener::bind("0.0.0.0:8080").await.unwrap();
     axum::serve(listener, app).await.unwrap();
